Gets to testcase 5

+++ src/FordFulkerson.java
@@ -28,7 +28,7 @@ public class FordFulkerson {
         }
     }
 
-    public double maxFlow() {
+    public int maxFlow() {
         return value;
     }
